refactor: 클래스 분리
RequestHandler에 통합돼 있는 역할을 HttpRequest와 HttpResponse로
분리했습니다.
